+++ MaceEvolve.Core/Models/NeuralNetworkStepInfo.cs
@@ -7,10 +7,10 @@ namespace MaceEvolve.Core.Models
     {
         public int NodeId { get; set; }
         public virtual NodeType NodeType { get; set; }
-        public double Bias { get; set; }
+        public float Bias { get; set; }
         public CreatureInput? CreatureInput { get; set; }
         public CreatureAction? CreatureAction { get; set; }
-        public double PreviousOutput { get; set; }
+        public float PreviousOutput { get; set; }
         public List<Connection> ConnectionsFrom { get; set; } = new List<Connection>();
         public List<Connection> ConnectionsTo { get; set; } = new List<Connection>();
         public List<Connection> Connections { get; set; } = new List<Connection>();
