+++ MaceEvolve.Core/Models/Node.cs
@@ -6,13 +6,13 @@ namespace MaceEvolve.Core.Models
     {
         #region Properties
         public NodeType NodeType { get; }
-        public double Bias { get; set; }
+        public float Bias { get; set; }
         public CreatureInput? CreatureInput { get; set; }
         public CreatureAction? CreatureAction { get; set; }
         #endregion
 
         #region Constructors
-        public Node(NodeType nodeType, double bias, CreatureInput? creatureInput = null, CreatureAction? creatureAction = null)
+        public Node(NodeType nodeType, float bias, CreatureInput? creatureInput = null, CreatureAction? creatureAction = null)
         {
             NodeType = nodeType;
             CreatureInput = creatureInput;
